backend/requirements.txt
fastapi
uvicorn[standard]
gridstatus
httpx
pandas
python-dotenv
passlib[bcrypt]
python-jose[cryptography]
motor
beanie
pydantic-settings
pydantic[email]
python-multipart